requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.1.0
Babel==2.9.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
ffmpeg-python==0.2.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-speech==2.31.0
google-generativeai==0.3.1
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.63.0rc1
gTTS==2.5.4
idna==3.10
Jinja2==3.1.5
llvmlite==0.44.0
magic-filter==1.0.12
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
propcache==0.3.0
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
python-dotenv==1.0.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
semantic-version==2.10.0
setuptools-rust==1.10.2
soundfile==0.13.1
srt==3.5.3
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
vosk==0.3.44
websockets==15.0
yarl==1.18.3
